+++ example/redash_one_node/main.tf
@@ -5,7 +5,7 @@ module "redash" {
   nomad_namespace   = "default"
 
   # redash
-  service         = "redash"
+  service_name    = "redash"
   host            = "127.0.0.1"
   port            = 5000
   container_image = "redash/redash:8.0.2.b37747"
@@ -13,7 +13,7 @@ module "redash" {
 }
 
 
-module "redis" {
+module "redash-redis" {
   source = "github.com/Skatteetaten/terraform-nomad-redis.git?ref=0.1.0"
 
   # nomad
@@ -21,7 +21,7 @@ module "redis" {
   nomad_namespace   = "default"
 
   # redis
-  service_name    = "redis"
+  service_name    = "redash-redis"
   host            = "127.0.0.1"
   port            = 6379
   container_image = "redis:3-alpine"
@@ -30,11 +30,10 @@ module "redis" {
     cpu    = 200
     memory = 128
   }
-
 }
 
 
-module "postgres" {
+module "redash-postgres" {
   source = "github.com/Skatteetaten/terraform-nomad-postgres.git?ref=0.4.1"
 
   # nomad
@@ -42,7 +41,7 @@ module "postgres" {
   nomad_namespace   = "default"
 
   # postgres
-  service_name    = "postgres"
+  service_name    = "redash-postgres"
   container_image = "postgres:12-alpine"
   container_port  = 5432
   vault_secret = {
@@ -59,3 +58,4 @@ module "postgres" {
   use_canary                      = false
   container_environment_variables = ["PGDATA=/var/lib/postgresql/data/"]
 }
+