+++ main.tf
@@ -1,27 +1,34 @@
 locals {
-  datacenters = join(",", var.nomad_datacenters)
+  datacenters              = join(",", var.nomad_datacenters)
+  redash_config_properties = join(" ; ", var.redash_config_properties)
 }
 
 data "template_file" "nomad_job_redash_server" {
   template = file("${path.module}/nomad/redash_server.hcl")
   vars = {
-    datacenters  = local.datacenters
-    namespace    = var.nomad_namespace
-    image        = var.container_image
-    service      = var.service
-    host         = var.host
-    port         = var.port
-    cpu          = var.resource.cpu
-    memory       = var.resource.memory
-    cpu_proxy    = var.resource_proxy.cpu
-    memory_proxy = var.resource_proxy.memory
-    use_canary   = var.use_canary
+    datacenters              = local.datacenters
+    namespace                = var.nomad_namespace
+    image                    = var.container_image
+    service_name             = var.service_name
+    host                     = var.host
+    port                     = var.port
+    cpu                      = var.resource.cpu
+    memory                   = var.resource.memory
+    cpu_proxy                = var.resource_proxy.cpu
+    memory_proxy             = var.resource_proxy.memory
+    use_canary               = var.use_canary
+    redash_config_properties = local.redash_config_properties
     # Redis
-    redis_service = var.redis_service.service
+    redis_service = var.redis_service.service_name
     redis_port    = var.redis_service.port
     # Postgres
-    postgres_service = var.postgres_service.service
+    postgres_service = var.postgres_service.service_name
     postgres_port    = var.postgres_service.port
+    //    # Trino
+    //    trino_service = var.trino_service.service
+    //    trino_port    = var.trino_service.port
+    # Datasource upstreams
+    upstreams = jsonencode(var.datasource_upstreams)
   }
 }
 
@@ -31,7 +38,7 @@ data "template_file" "nomad_job_redash_worker" {
     datacenters  = local.datacenters
     namespace    = var.nomad_namespace
     image        = var.container_image
-    service      = var.service
+    service_name = var.service_name
     host         = var.host
     port         = var.port
     cpu          = var.resource.cpu
@@ -40,11 +47,16 @@ data "template_file" "nomad_job_redash_worker" {
     memory_proxy = var.resource_proxy.memory
     use_canary   = var.use_canary
     # Redis
-    redis_service = var.redis_service.service
+    redis_service = var.redis_service.service_name
     redis_port    = var.redis_service.port
     # Postgres
-    postgres_service = var.postgres_service.service
+    postgres_service = var.postgres_service.service_name
     postgres_port    = var.postgres_service.port
+    //    # Trino
+    //    trino_service = var.trino_service.service
+    //    trino_port    = var.trino_service.port
+    # Datasource upstreams
+    upstreams = jsonencode(var.datasource_upstreams)
   }
 }
 
@@ -54,7 +66,7 @@ data "template_file" "nomad_job_redash_scheduler" {
     datacenters  = local.datacenters
     namespace    = var.nomad_namespace
     image        = var.container_image
-    service      = var.service
+    service_name = var.service_name
     host         = var.host
     port         = var.port
     cpu          = var.resource.cpu
@@ -63,11 +75,12 @@ data "template_file" "nomad_job_redash_scheduler" {
     memory_proxy = var.resource_proxy.memory
     use_canary   = var.use_canary
     # Redis
-    redis_service = var.redis_service.service
+    redis_service = var.redis_service.service_name
     redis_port    = var.redis_service.port
     # Postgres
-    postgres_service = var.postgres_service.service
+    postgres_service = var.postgres_service.service_name
     postgres_port    = var.postgres_service.port
+
   }
 }