+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 
 ## [0.1.0 UNRELEASED]
-
 ### Updated
 - Bump vagrant-hashistack ">= 0.10, < 0.11". #32
 - Removed unused files/folders and updated the folder structure. #32
 - Updated example/redash_one_node with upstream to postgres and redis, and dynamical variables. #34
+
+### Added
+- Added a new example/redash_trino_cluster #39
+- Added variable `redash_config_properties` for custom redash properties. This variable makes it possible to create a new data source dynamically. #39
+
+
+## [0.0.3]
 
 ### Added
 - update-box as a prereq for the template-example target in the makefile